phase 5n: uint8 KV cache matching Qualcomm's exact per-layer scales
Pinned our KV encodings to Qualcomm's metadata.yaml values exactly.
build_uint8_kv_overrides.py parses their AR1/CL512 metadata, emits
one override JSON per part (2/3/4) covering all past_kv inputs AND
present_kv outputs (same scale on both sides so KV can be concatenated
across decode steps without requantize - invariant from Phase 0).

Re-converted parts 2/3/4 with --quantization_overrides, re-quantized
with w8 + per-channel + CLE, re-bin-gen'd. DLC confirms bitwidth=8
for all 72 KV tensors per part with scales matching Qualcomm's.
Wrapper ONNXs rebuilt with uint8 KV ports.

Per-part probe: cos unchanged or slightly better (Part 4: 0.988321 ->
0.988884). uint16 -> uint8 lossless at observed KV distribution
(narrow range vs uint16's 32k ceiling).

End-to-end oracle:
  before (u16 KV): cos +0.282  decode ". what is gravity? keep"
  after  (u8  KV): cos +0.374  decode ". What is gravity? Keep"

Capitalization now matches Qualcomm's output, confirming the uint8
KV helps with numerical precision at positions where the model would
normally flip case (our narrower calibration was off in those
directions). Argmax agreement 28/46 (vs 29/46 before) - essentially
the same token-level match but better semantic alignment.

Structural match progress:
  Part 1: 778 MB = 778 MB  (exact)
  KV: uint8 symmetric offset=-128 with Qualcomm scales  (exact)
  cos/sin: full-dim vs Qualcomm half-dim  (Phase 5o next)
  Size: 4833 vs 3136 MB  (needs better w4 calibration -> AIMET 5p)

+++ scripts/build_specula_4b_wrappers.py
@@ -59,16 +59,18 @@ def part_io_spec(part: int) -> tuple[list[tuple], list[tuple], str]:
             ("position_ids_cos", "uint16", [1, 1, HEAD_DIM]),
             ("position_ids_sin", "uint16", [1, 1, HEAD_DIM]),
         ]
+        # Phase 5n: KV now uint8 (matching Qualcomm's genie bundle). Per-layer
+        # scales applied via --quantization_overrides at convert time.
         for li in range(s, e + 1):
-            items.append((past_name("key", li), "uint16",
+            items.append((past_name("key", li), "uint8",
                           [1, NUM_KV_HEADS, PAST, HEAD_DIM]))
-            items.append((past_name("value", li), "uint16",
+            items.append((past_name("value", li), "uint8",
                           [1, NUM_KV_HEADS, PAST, HEAD_DIM]))
         return items
 
     def decode_outputs(s: int, e: int) -> list[tuple]:
         return [
-            (present_name(kind, li), "uint16",
+            (present_name(kind, li), "uint8",
              [1, NUM_KV_HEADS, CTX, HEAD_DIM])
             for li in range(s, e + 1) for kind in ("key", "value")
         ]
